cosmetic changes to file structure putting `user` items in a `user_core` space
