+++ flask_security/__init__.py
@@ -12,7 +12,8 @@
 
 __version__ = '1.6.2'
 
-from .core import Security, RoleMixin, UserMixin, AnonymousUser, current_user
+from .core import Security, current_user
+from .user_core import RoleMixin, UserMixin, AnonymousUser
 from .datastore import SQLAlchemyUserDatastore, MongoEngineUserDatastore, PeeweeUserDatastore
 from .decorators import auth_token_required, http_auth_required, \
      login_required, roles_accepted, roles_required, auth_required
