+++ flask_security/core.py
@@ -19,6 +19,7 @@ from passlib.context import CryptContext
 from werkzeug.datastructures import ImmutableList
 from werkzeug.local import LocalProxy
 
+from user_core import AnonymousUser
 from .utils import config_value as cv, get_config, md5, url_for_security
 from .views import create_blueprint
 from .forms import LoginForm, ConfirmRegisterForm, RegisterForm, \
@@ -232,48 +233,6 @@ def _get_state(app, datastore, **kwargs):
 
 def _context_processor():
     return dict(url_for_security=url_for_security, security=_security)
-
-
-class RoleMixin(object):
-    """Mixin for `Role` model definitions"""
-    def __eq__(self, other):
-        return (self.name == other or \
-                self.name == getattr(other, 'name', None))
-
-    def __ne__(self, other):
-        return (self.name != other and
-                self.name != getattr(other, 'name', None))
-
-
-class UserMixin(BaseUserMixin):
-    """Mixin for `User` model definitions"""
-
-    def is_active(self):
-        """Returns `True` if the user is active."""
-        return self.active
-
-    def get_auth_token(self):
-        """Returns the user's authentication token."""
-        data = [str(self.id), md5(self.password)]
-        return _security.remember_token_serializer.dumps(data)
-
-    def has_role(self, role):
-        """Returns `True` if the user identifies with the specified role.
-
-        :param role: A role name or `Role` instance"""
-        return role in self.roles
-
-
-class AnonymousUser(AnonymousUserBase):
-    """AnonymousUser definition"""
-
-    def __init__(self):
-        super(AnonymousUser, self).__init__()
-        self.roles = ImmutableList()
-
-    def has_role(self, *args):
-        """Returns `False`"""
-        return False
 
 
 class _SecurityState(object):
