+++ flask_security/user_core/confirmable.py
@@ -14,9 +14,9 @@ from datetime import datetime
 from flask import current_app as app, request
 from werkzeug.local import LocalProxy
 
-from .utils import send_mail, md5, url_for_security, get_token_status,\
+from ..utils import send_mail, md5, url_for_security, get_token_status,\
      config_value
-from .signals import user_confirmed, confirm_instructions_sent
+from ..signals import user_confirmed, confirm_instructions_sent
 
 
 # Convenient references
